requirements.txt
Flask==1.1.2
Flask==2.2.2
Flask==2.2.3
keras==2.12.0
numpy==1.15.1
numpy==1.21.5
numpy==1.23.5
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.0.1
Pillow==9.3.0
Pillow==9.5.0
Requests==2.28.2
tensorflow==2.12.0
tensorflow_intel==2.12.0
